+++ resources/views/components/api-operator-chat.blade.php
@@ -1,0 +1,4 @@
+@if(\App\Support\ApiOperator::visibleOnRequest())
+    @include('partials.api-operator-widget')
+    @vite(['resources/js/api-operator-widget.js'])
+@endif